= Retail Fortress Siege = == Thanks Vice ==
''Core Information:''
This is the initial "retail-like" core implementation of the fortress sieges.
There are some new types, and also some new packets and static objects.
,,There may be a few lingering debug messages which are harmless, I'll take them out if we find any, but I did a quick sweet and didn't see any on first site.,,

''Datapack Infomation:''
The datapack contains the complete implementation of the Dragonspine fortress, but still requires some datapack contributions to complete the remaining fortresses as "close to retail" as possible.
So the Dragonspine fortress is enabled by default, while all the others are disabled since their "Suspicious Merchant" is not spawned. If the merchant is spawned the fortress will be "siegable" but if it is not, regardless, the sieges can be turned on and off from the administrator panel at any time without the need of the suspicious merchants.

Fortress geodata needs to be completed by the Geodata specialists please. Also speaking of geodata, castle geodata needs major improvements to aid against infamous issues.

'''''Quoting Vice:'''''
{{{
This has been testd for about 3-4 weeks by me, and a test run was done on Ahmed servers to make sure all started and worked properly.
Thx to Izzy for DP help, thx to all l2j team for help, and thx a lot to Ahmed to encouragement and help on testing . Hope u enjoy =]
}}}