TVT Event improvements by Merwllyra
This patch fixes following issues:
 * '''Exploitable: Teams can be forged by exiting and re-entering the competition'''
Players would check the team count and exit/enter the event to be able to choose which side they fight for, which made unbalanced teams for the event. This is mostly bad for small servers, but also affects High Rate servers. 
 * '''Bug: The teleporter teleports players in a fixed spot.''' A random radius around the spot should be used instead. 
When 20 players are teleported in the same spot, some of them will be stuck and won't be able to move.
 * '''Bug: Members of the same clan must use forced attack when playing against each other in TvT'''
 * '''Feature: Instead of "time to pass" values, use "time of day" values in TvT configuration'''
Most of the times it's better to have the events announced on the website/forum and the players will know when to come online to start playing. However, if you restart the server (due to a bug or maintenance) before the event, the timer would reset. This addresses the issue, and instead of an interval, a time of day is specified for each event.
 * '''Feature: Move TvT to the next phase by admin command'''
If for some reason an event should start before the announced time or should move to the next phase (due to teams full or whatever) a GM may use the `tvt_advance` command to skip the TvT event to the next phase.